requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
werkzeug==3.0.1
Flask-Mail==0.9.1
